feat: runtime config (JSON, load on startup)
Load ~/.config/qed/config.json (honoring $XDG_CONFIG_HOME) before
editor_init. The config.odin constants become mutable typed globals
(defaults); the file overrides them.

Always-materialized policy: a missing file is created with every key;
missing keys are written back with defaults so the file always shows
every knob; an invalid value keeps the default at runtime, is named on
the message line, and is left untouched in the file. A syntactically
broken file falls back to defaults without rewriting.

Colors are nested under a `theme` object with readable names
(foreground, background, syntax_keyword, git_added, ...). Keybinds bind
command name -> key string; Ctrl+letter is case-insensitive and shown
lowercase, Alt case is significant, "" means unbound (palette-only).

+++ src/editor.odin
@@ -241,7 +241,7 @@ editor_dispatch_mouse :: proc(editor: ^Editor, ev: tb2.Event) {
 		pos := editor_mouse_cursor(editor, int(ev.x), int(ev.y))
 		recent :=
 			editor.click_count > 0 &&
-			time.duration_milliseconds(time.tick_since(editor.last_click_tick)) <= DOUBLE_CLICK_MS &&
+			time.duration_milliseconds(time.tick_since(editor.last_click_tick)) <= f64(DOUBLE_CLICK_MS) &&
 			pos == editor.last_click_pos
 		editor.click_count = editor.click_count + 1 if recent else 1
 		if editor.click_count > 3 {
